Added files via upload
Alright so I tried to fix the continuous looping in the adding recipe portion of the project. Where basically it continuously asked for an ingrediants and the quantity. So I replaced the cases with if loops and basically got rid of the main menu void function and incorporated in the while loop where I was running the option to run the program but, the only it works is if you press enter 2 times quickly sorry it tried...